requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
aiofiles==24.1.0
nest-asyncio==1.5.8
aiohttp==3.9.1
beautifulsoup4==4.12.2
brotli==1.1.0
motor==3.5.1
pymongo==4.8.0
python-dotenv==1.0.0
httpx==0.25.2
cachetools==6.2.0
lxml==4.9.3
cssutils==2.8.0
google-generativeai==0.3.2